+++ src/cli.js
@@ -15,13 +15,13 @@ yargs
   .usage('Usage: frey <command> [options]')
   .example('frey backup -d ./envs/production', 'backup platform described in ./envs/production')
   .options({
-    projectdir: {
+    project_dir: {
       nargs: 1,
       type: 'string',
       describe: 'Directory that contains the Freyfile.toml. Frey will traverse upwards if empty. '
     },
     app: {
-      default: '{{{projectdir}}}|basename',
+      default: '{{{self.project_dir}}}|basename',
       nargs: 1,
       type: 'string',
       describe: "Name of application for which we're creating infrastructure"
@@ -120,7 +120,7 @@ const liftOff = new LiftOff({
 })
 
 liftOff.launch({
-  cwd: argv.projectdir
+  cwd: argv.project_dir
 }, env => {
   if (env.configBase === undefined) {
     const msg = 'Could not find a Freyfile.toml in current directory or upwards, or in project directory.'
@@ -128,7 +128,7 @@ liftOff.launch({
   }
 
   // Let LiftOff override the project dir
-  argv.projectdir = env.configBase
+  argv.project_dir = env.configBase
   const frey = new Frey(argv)
 
   // Bombs away