+++ test/scenario/gitless/run.sh
@@ -25,7 +25,7 @@ cp Freyfile.toml "${__sysTmpDir}/frey-gitless/"
   --verbose \
   --force-yes \
   --cfg-var "global.ssh.keysdir=${__sysTmpDir}/frey-gitless" \
-  --projectdir "${__sysTmpDir}/frey-gitless/" \
+  --project_dir "${__sysTmpDir}/frey-gitless/" \
   --bail
 
 rm -f Frey-residu* 2>&1 > /dev/null || true